+++ repos/ktor/common/src/commonMain/kotlin/dev/inmo/micro_utils/repos/ktor/common/ParametersNames.kt
@@ -1,0 +1,5 @@
+package dev.inmo.micro_utils.repos.ktor.common
+
+const val keyParameterName = "key"
+const val valueParameterName = "value"
+const val reversedParameterName = "reversed"