+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.4.0
+
+* `Repos`:
+    * `ReadOneToManyKeyValueRepo` got `keys` method with value parameter
+        * All implementations inside of this library has been updated
+    * `ReadStandardKeyValueRepo` got `keys` method with value parameter
+        * All implementations inside of this library has been updated
+    * New extensions `withMapper`
+
 ## 0.3.3
 
 * `Coroutines`: